LayoutTests/http/tests/security/popup-allowed-by-sandbox-when-allowed.html

<script>
if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.setCanOpenWindows(true);
    layoutTestController.setCloseRemainingWindowsWhenComplete(true);
}
</script>
<p>To run this test outside of DumpRenderTree, please disable your popup blocker!</p>
<iframe sandbox="allow-scripts allow-popups"
  src="data:text/html,
       <script>
       var win = window.open('data:text/html,<script>if (window.layoutTestController) layoutTestController.notifyDone();<\/script>', '_blank');
       alert(win ? 'PASS' : 'FAIL');
       </script>"
  ></iframe>
