LayoutTests/http/tests/security/popup-allowed-by-sandbox-when-allowed-expected.txt

ALERT: FAIL
To run this test outside of DumpRenderTree, please disable your popup blocker!
